role admin dan member

+++ routes/web.php
@@ -23,24 +23,24 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
     Route::get('/admin', [AdminController::class, 'index']);
     Route::get('/home', [HomeController::class, 'index'])->name('home');
-    
     // jenis_kendaraan
-    Route::get('/admin/jenis_kendaraan', [JenisController::class, 'index']);
-    Route::get('/admin/form_jenis/create', [JenisController::class, 'create']);
-    Route::post('/admin/jenis_kendaraan/store', [JenisController::class, 'store']);
-    Route::delete('/admin/jenis_kendaraan/destroy/{id}', [JenisController::class, 'destroy']);
-    Route::get('/admin/jenis_kendaraan/edit/{id}', [JenisController::class, 'edit']);
-    Route::put('/admin/jenis_kendaraan/update/{id}', [JenisController::class, 'update']);
-    Route::get('/admin/jenis_kendaraan/show/{id}', [JenisController::class, 'show']);
+    Route::get('/admin/jenis_kendaraan', [JenisController::class, 'index'])->middleware('auth','admin');
+    Route::get('/admin/form_jenis/create', [JenisController::class, 'create'])->middleware('auth','admin');
+    Route::post('/admin/jenis_kendaraan/store', [JenisController::class, 'store'])->middleware('auth','admin');
+    Route::delete('/admin/jenis_kendaraan/destroy/{id}', [JenisController::class, 'destroy'])->middleware('auth','admin');
+    Route::get('/admin/jenis_kendaraan/edit/{id}', [JenisController::class, 'edit'])->middleware('auth','admin');
+    Route::put('/admin/jenis_kendaraan/update/{id}', [JenisController::class, 'update'])->middleware('auth','admin');
+    Route::get('/admin/jenis_kendaraan/show/{id}', [JenisController::class, 'show'])->middleware('auth','admin');
+   
     
     // armada
-    Route::get('/admin/armada', [ArmadaController::class, 'index']);
-    Route::get('/admin/form_armada/create', [ArmadaController::class, 'create']);
-    Route::post('/admin/armada/store', [ArmadaController::class, 'store']);
-    Route::get('/admin/armada/edit/{id}', [ArmadaController::class, 'edit']);
-    Route::put('/admin/armada/update/{id}', [ArmadaController::class, 'update']);
-    Route::delete('/admin/armada/destroy/{id}', [ArmadaController::class, 'destroy']);
-    Route::get('/admin/armada/show/{id}', [ArmadaController::class, 'show']);
+    Route::get('/admin/armada', [ArmadaController::class, 'index'])->middleware('auth','admin');
+    Route::get('/admin/form_armada/create', [ArmadaController::class, 'create'])->middleware('auth','admin');
+    Route::post('/admin/armada/store', [ArmadaController::class, 'store'])->middleware('auth','admin');
+    Route::get('/admin/armada/edit/{id}', [ArmadaController::class, 'edit'])->middleware('auth','admin');
+    Route::put('/admin/armada/update/{id}', [ArmadaController::class, 'update'])->middleware('auth','admin');
+    Route::delete('/admin/armada/destroy/{id}', [ArmadaController::class, 'destroy'])->middleware('auth','admin');
+    Route::get('/admin/armada/show/{id}', [ArmadaController::class, 'show'])->middleware('auth','admin');
     
     // pembayaran
     Route::get('/admin/pembayaran', [PembayaranController::class, 'index']);
@@ -59,8 +59,8 @@ Route::middleware('auth')->group(function () {
     Route::put('/admin/peminjaman/update/{id}', [PeminjamanController::class, 'update']);
     Route::delete('/admin/peminjaman/destroy/{id}', [PeminjamanController::class, 'destroy']);
     Route::get('/admin/peminjaman/show/{id}', [PeminjamanController::class, 'show']);
-    Route::middleware('admin')->group(function () {
-});
+    //Route::middleware('admin')->group(function () {
+//});
 
 //Route::middleware('admin')->group(function () {
     // Tambahkan rute khusus untuk admin di sini jika diperlukan
